ChurnModel/training_pipeline_requirements.txt
# To generate this file, find the imports in your file
# Then see installed versions
# pip freeze|grep joblib
# > joblib==0.14.1
numpy==1.18.5
pandas==0.25.3
sklearn-pandas==1.7.0
lightgbm==2.3.0
joblib==0.14.1
azureml-sdk==1.19.0